Break pod filters out into their own file
